Modify the ER diagram's png and pdf copy
This modification have been made because an attribute called
customer_type has been added to CUSTOMER entity or table.

The attribute customer_type was added because ECG have different
packages for their customers(schools, factories and so on). So because
of this a new attribute have to be introduced to differentiate between
the customers and their package type. This affects their billing too.
This work does not go deep into how an ECG system works so this is okay
(this work is just a school work).